Implement different output types for LLM (BaseMessage) v Chat Model (str)
